Conf/okursoproj/requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
mysql-connector-python==8.0.33
mysqlclient==2.1.1
protobuf==3.20.3
PyMySQL==1.0.3
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
